Add files via upload
This is my first project using python, I used Python to explore data related to bike share systems for three major cities in the United States—Chicago, New York City, and Washington. The script takes in raw input to create an interactive experience in the terminal to present descriptive statistics.